requirements.txt
opencv_contrib_python==4.4.0.42
requests==2.26.0
pandas==1.2.3
numpy==1.21.2
Pillow==8.3.2
torch==1.9.0
torchvision==0.10.0
